# api/openapi.yaml
openapi: 3.0.3
info:
  title: PayHere API Documentation
  description: PayHere API Documentation
  version: 1.0.0
servers:
  - url: 'http://localhost:1202'
...
paths:
...
  /v1/users/signOut:
    post:
      tags:
        - user
      operationId: signOut
      summary: 로그 아웃
      description: |
        로그아웃 기능을 제공합니다.
        
        만료되지 않은 토큰의 경우 `token_blacklist`에 등록합니다.
        
        ### Error case
        
        - 인증이 실패할 경우, `Unauthorized (401)` 에러를 반환합니다.
        - 토큰이 이미 블랙리스트에 등록된 경우, `TokenBlacklistAlreadyExists (401)` 에러를 반환합니다.
        - 유저가 존재하지 않는 경우, `UserNotFound (401)` 에러를 반환합니다.
        - 서버 에러가 발생한 경우, `InternalServerError (500)` 에러를 반환합니다.

      security:
        - tokenAuth: []
      responses:
        204:
          description: OK
        401:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                Unauthorized:
                  $ref: "#/components/examples/Unauthorized"
                TokenBlacklistAlreadyExists:
                  $ref: "#/components/examples/TokenBlacklistAlreadyExists"
                UserNotFound:
                  $ref: "#/components/examples/UserNotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    tokenAuth:
      description: |
        토큰 인증,
        
        로그인 API를 통해 발급된 token을 전송합니다.
      type: http
      scheme: Bearer
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            UserAlreadyExists:
              $ref: "#/components/examples/InternalServerError"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ResponseMeta:
      type: object
      properties:
        code:
          type: integer
          description: 응답 상태 코드
          example: 200
        message:
          type: string
          description: 응답 메시지
          example: ok
  examples:
...
    Unauthorized:
      value:
        meta:
          code: 401
          message: Server failed to authenticate the request.
...
    TokenBlacklistAlreadyExists:
      value:
        meta:
          code: 401
          message: The specified token already exists in token blacklist.
    UserNotFound:
      value:
        meta:
          code: 404
          message: The specified user doesn't exist.
...
    InternalServerError:
      value:
        meta:
          code: 500
          message: The server encountered an internal error. Please retry the request.